bwl: Change read_acs_report()
Change the function read_acs_report() on all flavors of bwl:

1. On DWPAL, change it to fill the rank on the unified channels list,
instead of the preferred channels list.

2. On NL80211 the function called another function that copies the
supported channels into the preferred channels. Since this is not needed
anymore, change the function to return 'true'.

3. Remove old implementation that fills the preferred list manually, since it
is not needed anymore. The unified channels list is already filled on the
"referesh_radio_info()".

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>