bwl: Remove read_preferred_channels()
On some flavors of bwl (NL80211 and Dummy), the function read_acs_report()
was not implemented since reading ACS was not supported on these bwl flavors,
so there was no function to fill the "preferred channels list".
In addition, in case acs was not enabled on the platform, the preferred channels
list remained empty since the function read_acs_report() was not been called.
To solve these scenarios the function "read_preferred_channels()" has been
introduced, and filled the preferred channels list from the supported channels list.

Since now we have a unified channels list filled as part of the refresh_radio_info(),
we don't need this function anymore and it can be removed, so remove it.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>