agent: Update the function that built the unified channels list
The function unify_channels_list() receives the preferred and supported
channels list, and construct the unified channels from them (as well do
some ranking to preference conversion algorithm), and add it to the
cACTION_APMANAGER_CHANNELS_LIST_RESPONSE.
This is not needed anymore since the unified list is already constructed.
Change the function by removing the parts that construct the unify channels list.
Also, the function will now fill a given cChannelList tlvf container instead of
cACTION_APMANAGER_CHANNELS_LIST_RESPONSE to make it more generic
and be capable to fill the list on various messages.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>