mon_wlan_hal_dwpal: do not clear handled_clients
The generate-connected-events flow is a workaround to the scenario where
clients connected prior to the startup of prplmesh. As such, it will be
done only once (in several iterations but once completes, will not be
repeated).

Keep the handled_clients container and the completed_vaps container to
prevent duplication of clients associsation events.

The clients are now inserted to the container only for VAPs that are not
part of the completed_vaps.

PPM-1409.

Signed-off-by: Adam Dov <adov@maxlinear.com>