transport: remove promiscuous mode for not-multiap mode
The promiscuous mode is required to pass packets destined to AL_MAC
which might be different than the interfaces' HW address.

However, in not-multiap this is not required, and on the other hand, it
may interfere with applications running on the bridge interface (e.g.
DNAT).

Read the management mode and set promiscuous mode only for non
not-multiap configuration.

Signed-off-by: Adam Dov <adov@maxlinear.com>