boardfarm: add maximum_tcp_throughtput test
After trying to rely on the tlvs to switch the channel on glinet to
36 to fullfill the test requirements and ending up with no effect (
the channel stayed on 48) it was decided that for now the channel
requirements should be skipped, along with the channel bandwidth
The channel_selection test is also failing on the glinet
On the throughtput speed validation, the retrieved results are way
under the pass criteria.
Max download 2.4GHz > 28.8 vs 90
Max upload 2.4GHz > 38.5 vs 90
Max download 5GHz > 87.9 vs 504
Max upload 5GHz > 88.1 vs 504
For now the testbed is still not on its final location, so that
should weight in on the values, even thought not by much.
I was decided that for we're only gathering base line values
instead of treating it as a failing factor.
About the data on the cpu usage, since there's no test requirements
about it, and after reviewing it with Oscar, the feedback was that
it was only meant to have some additional info to grab on to in
case there's a need to investigate some kind misbehaviour

Signed-off-by: David <davidmgafonso@gmail.com>