bwl: Change channels list type
Change the channels list and its internal bw_info list to
std::map instead of a std::unordered_map.
The reason for this change is that storing the information in
unordered_map is causing the channels list to be printed in
a messy way which makes it difficult to debug.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>