agent: Add filter
The function get_channel_preferences_from_channels_list()
on the son_slave transforms the channel preferences to an
easy-to-use container for filling the channel preference report.
To do it, the function iterates over the operating classes on the
operating classes table in the wireless utils and checks the
status of the channel.
When the channel is on the 2.4 GHz band, some checks are
irrelevant such as DFS or overlap channel that is not supported.
For that cases add a filter the do these check only for the 5 GHz
band channel.

Also, add/fix prints.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>