agent: Remove channels list request from coordinated cac task
The coordinated cac task has implemented a code to pull the channels list from the AP.
The reason for it is that the unified channels list has been updated on the AgentDB only
on manual request, so the task requested the list if it was empty.
Since now it is promised that the channels list will be filled as soon as the AP Manager
has joined, the code that is responsible for taking the list can be removed.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>