controller: channel preference report handling in controller
Channel preference flow is not fully adapted (PPM-201).

To pass certification tests, dummy reply is prepared and send to
cmdu_tx_header.

The check of certification mode was blocking sending ACK message in case
of it is disabled.

ACK message and handler for CAC Status report is carried above
certification check.

https://jira.prplfoundation.org/browse/PPM-622
--
Signed-off-by: Cihangir Odabas <codabas@assia-inc.com>