controller: handle_control_message: check hostap for ZERO_MAC
The `hostap_mac` parameter is a result of convertion of
a MAC address to a string.

An "empty" MAC is 00:00:00:00:00:00 and after conversion
to a string ir would equal to "00:00:00:00:00:00".

The existing check for being empty is equivalent to
comparing a string to "".

Use the correct "empty" value for this check.

PPM-1060

Signed-off-by: Vitalii Komisarenko <vitalii.komisarenko@mind.be>