agent: Fix printing bug
In the commit cf096bd14d73176031a1bd649f77503874c416af I mistakely removed
the '+' operator from the stream output.
I forgot that only on easylogging we can print characters as integers and in a regular
stream it will not work so we must cast it to an integer.
Revert the '+' operator removal from the stream printing.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>