bpl_db: call uci_save on uci_delete_section
Removing clients from the run-time DB did not remove them from the UCI
hard-file; this causes issues due to the desynchronization of the two.

Although the library's documentation does not require this, calling the
save operation on the config solves this issue.

Note that calling the save operation does not commit the changes made
to the hard-file but only stores it in the temporary memory and does
not replace the commit operation.

PPM-1503

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>