tools: deploy_firmware: Small changes needed for correct work
Unpacking RDKB rootfs from tarball via tar
command generates a lot of output because
verbose key used. Need to remove it.

There is present some calls of method
check_serial_type() without parameters it is
causing an issue with broken logic of the script
because method does not work properly.

In method read_remote_rdkb_version() version
variable should be filled up anyway because
we can get situation when we are trying to flash
new board first time or RDKB rootfs on the board
suddenly brocken and file with version does not
exist.

PPM-1049

Signed-off-by: Vladyslav Tupikin <v.tupikin@inango-systems.com>