cient_steering_taski: set duration of client steering
According to PPM-991. Save  the timestamp when a disassociation event is received.
When the association event is received, the duration is recorded and the disassociation
timestamp is cleared. If the disassociation timestamp was not set, the duration is considered to be 0.
If the duration is smaller than some compile-time defined threshold, e.g. 10s,
it is considered a steering event originated by the station. duration == Time Taken.

PPM-991
--
Signed-off-by: Krystyna Oliinyk <k.oliinyk@inango-systems.com>