framework: remove unused #include <bpl/bpl_err.h>
If `<bpl/bpl_err.h>` is included by a file, one of symbols/
macros/etc. must be used.

A script (see [1]) was used to check it. It printed places
where `bpl_err.h` was included and all lines that mention
one of the entities defined by `bpl_err.h`.

If in a file there is only `#include` statement without
any other lines printed, the `#include` can be removed.

There are also several files where `eErrorCode` was used
but `bpl_err.h` was not included. It means that `bpl_err.h`
was included indirectly. This problem was NOT addressed
by this commit.

[1]:
find agent/ controller/ common/ framework/ \
    -name '*.cpp' -or -name '*.h' \
    | xargs grep -nP --color \
    'include.*bpl_err.h|BPL_ERROR_STRING_LEN|GENERATE_ERROR|ERROR_CODE|eErrorCode'

Signed-off-by: Vitalii Komisarenko <vitalii.komisarenko@mind.be>