controller: control_message: convert hostap_mac to sMacAddr
`hostap_mac` used to be a `std::string` value.
With the transition to `sMacAddr` variables, the code of
`handle_cmdu_control_message` became full of
`tlvf::mac_from_string` statement.
It makes code harder to read.

Transform `hostap_mac` into a pair of variable:
* `radio_mac` (contains `sMacAddr`)
* `radio_mac_str` (contains `std::string`)

In some places `radio_mac` was used as a local variable.
It contains `sMacAddr` version of `hostap_mac`, so there is
no conflict.

Do not change the handler of
`ACTION_CONTROL_CLIENT_RX_RSSI_MEASUREMENT_RESPONSE` because
it uses assignments to the radio mac (called `ap_mac` there).
It would require to synchronize `std::string` and `sMacAddr`
versions of the radio mac.

Signed-off-by: Vitalii Komisarenko <vitalii.komisarenko@mind.be>