agent: platform manager: do not print empty country code
If 'country_code' is not set, it contains two `\0` characters.
If printed as is, it will confuse `grep`.

By default `grep` checks the beginning of a file and if it
finds non-printable characters, the file will be considered
binary and skipped.
`grep -a` disables skipping such files, but it easy to forget
`-a` part.

Print `(not set)` if country code is not set.

Signed-off-by: Vitalii Komisarenko <vitalii.komisarenko@mind.be>