agent: Change Agent configuration struct
In this MR, the son_slave thread is being unified into a single thread.
Therefore, it is necessary to reorganize the agent configuration struct
to contain common and separate radio configurations.
On a future commit, the main file will fill this struct for all radios and
pass it to the class constructor.

Note that "radio_identifier" does not exist in the new struct. The reason
is that this field is not being used anywhere, and will be removed
completely on a future commit.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>