bcl: Rename modules names and uds files names
Until now there were two (or more) son_slave, so each son_slave name
was "beerocks_agent_" + <iface name>.
Since from this MR there is going to be only one, it should be just
"beerocks_agent" with not interface name suffix.
The problem is that this also is the name of the thread that contains
the backhaul and the platform.
To solve this out, add a new name for the backhaul/platform and replace it.

As part of this MR, changed the uds names to be more correlated with one another.
Also took this opportunity to convert the module's names and uds names to constexpr.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>