agent: Create a single Agent thread
On the agent_main file, run a single agent thread instead of multiple,
and change the fill_son_slave_config() function to fill the configuration
of the new form and was introduced on a previous commit.

Also, change the name of the file to "agent_main" instead of "slave_main".

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>