agent: Two son_slaves contexts in one Agent
Move all the son_slave members (including the FSM state) associated
with specific radio to sManagedRadio as a value to fronthaul interface
keyed map "m_radio_managers".

To every function that needs the fronthaul interface, add the interface
as a function argument.

On the follow-up commits the CPP file will change accordingly.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>