agent: Handle autoconfiguration renew
When a vendor-specific message gets to the backhaul manager from the Controller,
the backhaul just sends it to all son_slaves, and each son_slave filter it if the message
is not destined for it by comparing the mac address of the radio in the beerocks
message header to the local radio mac.
The standard message does not work like this. If a message is destined to a specific
radio or bssid or band, it is mentioned explicitly in the message, otherwise, it is
a message destined to all radios.
In the autoconfiguration renew message, the standard mentions the band.
To find which radio context should we treat, find which radio on the database
supports the required band.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>