agent: Handle messages with no specific destination
Some of the standard messages are sent generally to the agent
with no specific radio/bssid destination.
In that case iterate over, all the radio context do what needs to be done.

In the multi_ap_policy_config_request, it is possible to see a benefit of
the unified Agent architecture: Instead of sending bss error response
from each bss context, we can send it once.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>