agent: find the correct fronthaul socket to forward the message
The backhaul manager forwards messages to the son_slave in multicast.
In some cases, we need that each radio context will forward the message
to its radio context fronthaul socket.
Add a code that matches the received socket to the context radio so we
will be able to forward the message to the right fronthaul socket.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>