bpl: add uci flag to control clients measurements
Preparative Commit.
Currently, even if the user disables the steering and roaming features,
the clients' measurements will still be read from hostapd.
This, adds a lot of unnecessary reads from hostapd using dwpal.

To reduce this, it is important to add a capability to measure only
selected clients, or disable the measurements entirely.

Added new configuration flag to control if whether the monitor will
measure clients and which (none/all/selected):
0 - disable-all
1 - enable-all (default value)
2 - measure-only-selected-clients

PPM-1519.

Signed-off-by: Adam Dov <adov@maxlinear.com>