monitor_db: add the measurements mode flag
Preparative Commit.
The configuration can be read from UCI only once and used throughout the
monitor lifetime since any change requires prplmesh restart.

Add parameter to the monitor_db to store the read configuration.
Add APIs to get and set the parameter.
Add enum, equivalent to the BPL enum that can be used in the monitor_db
header without the need to include the BPL in the header.

PPM-1519.

Signed-off-by: Adam Dov <adov@maxlinear.com>