monitor_db: add sta measurements enable configuration
Preparative Commit.
In the "only_selected_clients" mode the configuration of which clients
to measure might be set before the clients' connection, and since the
measurements are done many times and we don't want to look up that
client in another container every-time, it is better to configure for
each client if it is to be measured or not and update that flag as
needed.

Add flag to enable/disable measurements for a client.
Add APIs to set/get the flag.

PPM-1519.

Signed-off-by: Adam Dov <adov@maxlinear.com>