agent: Backhaul forward to a single son_slave socket
Since there is now only one Agent thread, there is no point that the backhaul
manager will forward a message from the controller to all son_slave sockets,
since eventually they will be handled at the same thread anyway.

The function "handle_1905_1_message()" has a "forward_to" output argument
which lets the function choose to which son_slave to forward the message.
This argument is not being used, so remove it, and remove the code on the
caller function "handle_cmdu_from_broker()" that depends on it.

Replace the code that forwards the message to all son_slaves, with code that
forwards the message only to the first son_slave.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>