bwl: nl80211: add get_tx_power_limit_dbm()
This is a preparatory commit.

In a follow-up commit, the new method will be used to get the actual
tx power in refresh_radio_info().

Note that since we're converting mBm to an (unsigned) integer dBm
value, we're going to lose some precision.

This is however consistent with how the tx power is stored and
calculated in other implementations.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>