tlvf: Add interface to the Monitor join message
Preparative commit.
The AP manager and the Monitor threads are the only thread that
opens a socket to the Agent (son_slaves).
The AP manager reports on the first message it sends, to which interface
it belongs but the monitor does not.
Now that we have only one Agent thread it is required to add it also to
the Monitor first message so the Agent thread will be able to bind the
Monitor socket which through it received the message to the correct
interface.

Adds interface name parameter to MONITOR_JOINED_NOTIFICATION.

Reading the new parameter will be added in a future commit.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>