agent: General use of radio_manager
This commit is enormous, but all the changes in it have the same
principle that repeats over and over again:
Use the newly introduced "m_radio_manager" and its data members
to understand with which radio we are working, find the fronthaul
interface name, and use it to access the database on the right radio
instance.

This is a find-replace style commit. There is no new logic in it.
The changes take place on the class constructor & destructor,
all overridden functions (except the "handle_cmdu" which be
handled separately), FSM, vendor-specific message handlers,
and helper functions.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>