agent: Match socket to fronthaul interface on Handle CMDU
When a message is received on the Agent, all the handlers (except control)
need to know to which radio context it belongs. The messages from the
Controller do not need it, since the radio context can be understood from
the message itself (if there is even any). On VS message it is on the beerocks
message header.

To know to which radio context the message belongs, find to which of the
radio contexts sockets the incomings socket matches.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>