agent: Update control message handlers
On the control_ieee1905_1_message handler, remove all the checks
before the message type switch case since they are not relevant
anymore. As being said in the previous commit message, the 1905.1
Control messages have an internal field that implies the interface name.

Note that the zwdfs filter has been removed. This one is actually needed.
Solution for that will be provided on a separate commit.

On the VS control message handler, use the radio mac in the beerocks
message header to find the radio context.
Add the zwdfs filter that was missing, and remove the master socket
check since there are not relevant anymore because the message could
be sent from a different radio context master socket.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>