agent: Update fronthaul first message handlers
Now when the there is only a single agent thread, when a message is received
on the Agent, and handle_cmdu function tries to find to which of the fronthaul
interface the incoming socket matches, so it could pass the interface to the
handler function.
On the Fronthaul (ap manager and monitor), this will not work on the first
message they send.
The first messages they send are used to bind the incoming socket to the
interface name, so before getting these messages the fronthaul interface
will be passed empty to the handler function and will be dropped there.

To solve this, move the first message handler outside of the message
switch-case and before the fronthaul empty check. This will allow saving
the socket on the radio context so the other messages could be handled.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>