agent: Update replies to the controller
On some 1905 message handlers, we need to send back a message to the controller.
Since it is not important from which radio context controller socket the message
should be sent, choose the socket on the first radio context.
After PPM-351 there will be a single socket to the controller which does not belong
to a radio context anyway, and after the unified Agent will be completed, there will
be no socket at all, the message will be sent to the broker.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>