agent: Agent ZWDFS flow
A ZWDFS radio that exists on the platform is unique. The controller
must not be aware of its existence, and therefore it should not reply
to Controller Messages or action requests.

Since there are 1905 messages which are general and do not have
a specific radio to act on it, we are using iteration on all the radio
context to forward the message for example.
This could be a problem since the ZWDFS radio is part of the radio
contexts list.

There are two ways to solve this:
1. Add a ZWDFS filter for every 1905 message that iterates over
the radio context.
2. Move away the ZWDFS from the radio contexts list and put it
somewhere else.

Option 1 is not a good option, since adding a ZWDFS filter to each
place where iterating on the radio contexts is something easy to
forget, and we can't expect developers to remember.
Option 2 is preferred since the ZWDFS has finite functionality,
and it is much easier to use a side radio context when handling
ZWDFS flow.

This commit changes the following:
1. Create a unique_ptr to hold the ZWDFS context. When a ZWDFS radio
is discovered move its radio context to the unique_ptr, and erase it
from the radio contexts list.
2. Create a separate call to the ZWDFS FSM on the work function.
3. On the code that is part of the ZWDFS flow, use the side ZWDFS
radio context, instead of the context from the regular contexts list.

To do all of that, wrap the radio manager map with class that allows
do in more easly.
Now everytime iterating on the radios, the ZWDFS radio will be excluded.
It will be accessable only by if using the access operator with the
ZWDFS interface or by using "get_zwdfs()".

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>