tests: Update higher-layer data payload trigger test
Now, all radios in the boardfarm environment point to the same log file
(beerocks_agent.log). The "higher-layer data payload trigger" check that the
message is not handled on both radios by checking in the log file. Since it is
the same logfile, it will find it on both of them.
Remove the checks that test if both radios received the message since it is
irrelevant to do it now.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>