bcl: Fix wrong return value in send()
The function SocketConnectionImpl::send() return the number of bytes sent,
or -1 on error.

CmduPeer::send_cmdu() returns bool, but on the return statement it calls to
SocketConnectionImpl::send(). This means the return value of the function is
not correct.
Fix the return statement so it will return a correct return value.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>