bwl: Fix unitialized mcs_set value
On dummy, the values of ht_mcs_set and the vht_mcs_set are not initialized,
and not being set.
More over it is defined as "std::basic_string<uint8_t>" which has initial size of 0.
On the flow of drill these parameters down to the Controller, std::copy_n is being
used, on N size of beerocks::message::HT_MCS_SET_SIZE. On Dummy, this ends
with the copying of garbage into the TLV sent to the Controller, which with luck,
was zero until now.

Change the type of these fields to arrays with the fixed size which it should have,
and add set it to zeros on the BWL Dummy implementation.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>