controller: db: dm: LinkMetricsRequestInterval is added in NBAPI
Controller.Configuration is expanded with LinkMetricsRequestInterval.

This attribute holds Link metrics request interval in seconds.

It is initialized and filled from BPL, then is set to data model object.

Any changes which are triggered from NBAPI are stored within BPL
and are hold as persistently.

Also, any changes in this configuration model immediately
affects to flow in run time.

https://jira.prplfoundation.org/browse/PPM-1140
--
Signed-off-by: Cihangir Odabas <codabas@assia-inc.com>