framework: WSC: set default authentication and encryption type
The authentication and encryption type we set in our encrypted
settings (in the M2) were not initialized. As a result, in cases where
we don't fill them in (when we're supposed to teardown the radio) they
would have random values.

Instead of setting them only in the teardown case, initialize them to
default values.

While at it, set the default value for the bss_type, to avoid sending
uninitialized data in all cases. Since setting the teardown bit is a
requirement, still set in when building the M2 in case the default
value changes in the future.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>