controller: db: make db class a friend of station class
Preparatory commit.

In some cases station member variables can be coupled with
the DB configuration or with the legacy node structure.

In theses cases in is safer to make such member variables
private.
To keep the station class simple and to avoid saving DB
context in each station, the access will be done through
the DB class.

Make the DB class a friend of the station class to allow this.

PPM-1058

Signed-off-by: Vitalii Komisarenko <vitalii.komisarenko@mind.be>