ci: cppcheck: run on agent folder in gitlab-ci
Once there was a high number of cppcheck issues in the agent
code, so it was chosen to run this tool only on `common`,
`framework` and `controller`.

Enable `cppcheck` on `agent` folder.

Whitelist `selfAssignment` warning in `platform_manager.cpp`.
The code that causes this issue implements a workaround
to securely remove passwords in memory.

Whitelist `syntaxError` error because it is clearly a false
positive. If there was a syntax error the code would not
compile anyway.

Do not provide exact line numbers to be whitelisted because
they are likely to change.

PPM-951

Signed-off-by: Vitalii Komisarenko <vitalii.komisarenko@mind.be>