controller: db: update collection interval object
This attribute (CollectionInterval) needs to be updated:

* When a device (agent) is connected
* When LinkMetricsInterval is changed

Update method needs to be called in such cases.

For now, any configuration change triggers
event_configuration_changed event.

In this event, all configurations are copied to master config.
Also, CollectionInterval is updated.
Even if the data is same, updating mechanism is triggered.
Due to same values, working behavior will not be affected.

https://jira.prplfoundation.org/browse/PPM-463
--
Signed-off-by: Cihangir Odabas <codabas@assia-inc.com>