bcl: Add EventLoop Thread utility
Some functions that are currently run on the same thread need to be
separated to different threads. To make it possible, each thread need to
hold a separate EventLoop and Timer Manager instances.
In current implementation, for example fronthual_manager, the main
function creates the EventLoop instance and pass it to the different
classes for usage. With this implementation it is impossible to separate
the functionality into threads.

This commit incroduces EventLoopThread class that helps create a thread
with EventLoop.
In later commits, Monitor will be refactored to derive from this class.

PPM-1451

Signed-off-by: Semyon Barenboym <simonbaren@gmail.com>