tlvf: Change Platform Register request and response messages
Remove the interface name from the Platform registration message,
since from now there will be only one Agent socket in the platform manager.

In addition, remove the "valid" field from the response message.
The "valid" field is used for telling the Agent if the reading of the platform
configuration was successful.
On a follow-up commit, the function that fills the platform settings on the
platform manager will be moved to the Agent context, so the "valid" field
will not have a use anymore.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>