agent: Add global FSM
Create global states and an FSM function. These states will not be radio-specific
and instead be general for the Agent.
Currently, add only the platform registration state, and in a future MR also the
backhaul connection states will be moved there too.
Notice that a LOAD_PLATFORM_CONFIGURATION was added. The reason for that
is that on a follow-up commit the function "fill_platform_settings" on the platform
manager will be moved to the Agent context.
The idea behind it is to let the platform manager be responsible only for blocking
operations like setting configuration which do something in the platform. That is
because the code going in direction of doing bpl reading from anywhere and not
exclusively on the platform manager.

The use of the new states and FSM will be introduced in a follow-up commit.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>