agent: Move fill_platform_settings() to the Agent
Move the function fill_platform_settings() to the Agent.
Update the function so it will read the configuration that is radio specific
for each radio separately, but the common configuration is read once.
The main reason for doing this is that this is a relatively quick operation,
and regardless of the time it takes to read the settings, at that time, the
Agent does not do anything important anyway. Doing this in the Agent
allows to optimize performance a little bit, and remove some code from
the platform.
The function will be used in the Agent context on the next commit.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>