agent: Update Agent FSM
Add the code that split the radio contexted FSM to radio FSM and global Agent FSM:
1. Update the work() function to run first the Agent FSM and then the radio FSM.
2. Update the global Agent FSM state to work properly in separation from the radio context.
In these states, use the newly added do_on_each_radio_manager() to do some changes in
the radio that must be in that place.
Other radio changes that could be done closer to where they are used in the Radio FSM
moved to there.
3. Add the implementation of the new state LOAD_PLATFORM_CONFIGURATION.
4. Remove the platform registration connection retry mechanism since it is pretty redundant,
and I don't even remember when seeing such an error in the past few years.
5. Update the slave_reset() function to reset the FSMs properly.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>