Agent: Update socket related function on platform manager
Update socket related function on platform manager:
1. Remove the function add_slave_socket() since it has no content now that there
is only a single Agent file descriptor. Instead, update the file descriptor on the
Agent registration handler.
2. Same for the del_slave_socket() function, only used in the disconnect handler,
so remove the socket from there.
3. The function send_cmdu_safe() is renamed to send_cmdu_to_agent_safe(),
since it is only being used to send messages to the Agent.
4. get_slave_socket_from_hostap_iface_name() function removed since it serve
not purpuse now.
5. get_backhaul_socket() renamed to get_agent_socket() since this is what it
actually did - return socket to the Agent. Previously, it was able to return the
backhaul manager socket if the agent socket has not been found, but the
messages that need to be sent which are using this function are handled
only on the Agent anyway!

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>