controller: db: remove node key related code
Nodes are usually identified by their MAC address, represented
as a string.

A single MAC address is not enough to identify a radio,
a combination of AL MAC and radio UID is needed.
In the legacy node structure it is called "key".

If a node has a non-empty key, it basically has two
identifiers in the DB - its MAC and the key.

The last use-case of node keys by the DB users was removed
in the previous commit.

Internally, keys are used in `add_node` and `remove_node` -
basically to create/remove a node twice.

Remove code using node keys.
Remove `get_node_key` method that calculates a key.
Remove `radio_identifier` parameter of node that enables
calculating node key.
Remove `radio_identifier` from `add_node` method.

Signed-off-by: Vitalii Komisarenko <vitalii.komisarenko@mind.be>