framework: BPL: use a different variable name for bridge-ugw
The UGW version of libbridge was initially called "libbridge". At some
point it was renamed to "libbridge-ugw" in prplWrt, to avoid it from
conflicting with prplWrt's base "libbridge".

The MaxLinear builds use an older prplWrt tree, that still names the
UGW variant "libbridge".

To make it possible to build prplMesh with older trees,
d2d74fecb1fc15458e344ea3137c2919528b8b88 added a way to build with the
old name ("libbridge") if "libbridge-ugw" was not found.

However, this introduced an issue when configuring for a second time
with the old name ("libbridge"). By default, find_path caches the
result. On a rebuild, LIBBRIDGE_INCLUDE_DIR exists (it points to the
directory where "libbridge" is located), so we would fall in the first
part of the if statement. This part would try to find
"libbridge-ugw.a" in a directory that contains only "libbridge.a", so
it fails.

To fix it, just use a different variable name for libbridge-ugw's
include directory.

This should fix the intermittent failures of the external CI's UGW
builds where, for some reason, prplMesh sometimes is reconfigured, and
sometimes not.

While at it, add a comment to explain why the second part of the if
statement is necessary.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>