ap_wlan: add API to clear generate clients events progress
The agent propagates client connection events to the controller, if
there is a controller connected. To generate these events after
connecting to the controller (which can happen after a client already
connected) the agent triggers the flow to generate these events for any
connected client, right after the slave-join response.

To prevent duplicate events for already handled clients, the monitor and
ap-manager maintain internally the list of handled clients and handled
VAPs.

It was seen that when a client connected (and was added to
"handled_clients") and only then, connection to the controller was
established, that that client was not recognized by the controller and
did not appear in the bml connection map.

To resolve this, we need to clear the progress of the
generate-connected-clients-events before starting it.

Preparative commit.

Add API to clear generate clients' events progress in the ap_wlan_hal.

PPM-1566.

Signed-off-by: Adam Dov <adov@maxlinear.com>