agent: channel scan task: send scan to driver
Currently there is no validation for channels that are sent to the
driver from the agent's channel scan task.
This causes issues where the driver cannot complete the scan and thus
it results as a timeout error as described in PPM-1434.

To resolve this issue add validation for the channels to be scanned
before they are sent to the driver.
If a channel is not supported by the radio, send in its report the
appropriate.

PPM-1434.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>