bwl: initilize scan params structure
To support newer releases of the dwpal API we've created two structures
for the background params. we try to read the extened params if
possible.
But, if the dwpal API uses the legacy params, the internal values of
the extended structure stay uninitilized.

To resolve the issue of uninitilized scan params, we need to create a
default initialization for the structures' internal values.

PPM-1434.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>