test: Fix ClientAssociation test
Now the Agent uses standard messages internal BSSID fields to find out to which radio
a message has been sent.
The clientAssociation test sends CLIENT_ASSOCIATION_CONTROL_REQUEST_MESSAGE
to the Agent but on the BSSID field, it puts a radio mac instead of a BSSID mac.
This leads to a failure in finding the BSSID on the AgentDB.
Change the test to send in the TLV a BSSID mac instead of a radio mac.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>