framework: tlvf: add conditional elements keys
This is a preparatory commit.

These keys will be used to specify under which condition an element is
present.

The format is meant to be the following:

my_element:
  _condition:
    _field_name: my_presence_field
    _field_value: MY_ELEMENT_PRESENT

"_field_name" is the name of the presence field.

"_field_value" is the value it should have for the element to be
present.

The original idea was to introduce a condition in a single key (for
example "my_presence_field == MY_ELEMENT_PRESENT". However, having a
separate name and value will allow us to also set the presence field
when we set the element (for an outgoing TLV).

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>