ci: builders: remove Ubuntu Xenial
Ubuntu Xenial has python3.5, and [2591] requires python 3.6+ (for type
hints).

We either have to bring in a more recent python version to keep on
building with it, or remove it.

After a discussion in the team leader meeting, it was decided that we
would keep only one Ubuntu builder image so that we only have one to
maintain, so remove it and keep only Ubuntu Bionic.

Note that there is another builder image that uses Ubuntu
Xenial ('tools/docker/builder/rdkb/toolchain/Dockerfile', for
MaxLinear's rdk-b), so it will most likely also have to be updated or
removed.

[2591]: https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/merge_requests/2591

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>